+++ lib/commercetools-api/src/Models/Cart/ShippingBuilder.php
@@ -105,7 +105,7 @@ final class ShippingBuilder implements Builder
     }
 
     /**
-     * <p>Custom Fields of Shipping.</p>
+     * <p>Custom Fields of Shipping with <code>Multiple</code> <a href="ctp:api:type:ShippingMode">ShippingMode</a>.</p>
      *
 
      * @return null|CustomFields