+++ changes.md
@@ -3,6 +3,12 @@
 <details>
 <summary>Added Property(s)</summary>
 
+- added property `shippingKey` to type `Cart`
+- added property `shippingCustomFields` to type `Cart`
+- added property `shippingKey` to type `StagedOrder`
+- added property `shippingCustomFields` to type `StagedOrder`
+- added property `shippingKey` to type `Order`
+- added property `shippingCustomFields` to type `Order`
 - added property `key` to type `TaxRate`
 - added property `key` to type `TaxRateDraft`
 - added property `taxRateKey` to type `TaxCategoryRemoveTaxRateAction`